의상 조회 화면의 조회 프로세스를 수정하였습니다.
