+++ frontend/src/Presenter/Screens/Main/components/Category/CategoryContainer.jsx
@@ -1,19 +1,40 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
+import { fetchCategoryList } from "../../../../../httpRequest";
 import CategoryPresenter from "./CategoryPresenter";
 
-const CategoryContainer = ({ clothList, secondFilterIdx, page, setMaxPage }) => {
-  const categoryList = ["상의", "하의", "아우터", "모자", "신발", "가방"];
+const CategoryContainer = ({ secondFilter, page, setMaxPage }) => {
+  const [categoryClothList, setCategoryClothList] = useState([]);
+  const [nowPageList, setNowPageList] = useState([]);
 
-  // 의상 리스트 중 카테고리에 맞는 의상 리스트
-  const filteredClothList = clothList.filter((c) => c.type === categoryList[secondFilterIdx]);
+  useEffect(() => {
+    fetchCategoryList(secondFilter).then((r) => {
+      setCategoryClothList(r);
+      setNowPageList(r.slice((page - 1) * 10, page * 10));
 
-  // 카테고리에 맞는 의상 리스트 중 현제 페이지에서 보고 있는 의상 리스트
-  const nowPageClothList = filteredClothList.slice((page - 1) * 10, page * 10);
+      // 마지막 페이지 번호 설정
+      setMaxPage(Math.ceil(r.length / 10));
+    });
+  }, [page, secondFilter]);
 
-  // 마지막 페이지 번호 설정
-  setMaxPage(Math.ceil(filteredClothList.length / 10));
-
-  return <CategoryPresenter clothList={nowPageClothList} />;
+  return <CategoryPresenter clothList={nowPageList} />;
 };
 
 export default CategoryContainer;
+
+// const SeasonContainer = ({ secondFilterIdx, page, setMaxPage }) => {
+//   const [seasonClothList, setSeasonClothList] = useState([]);
+//   const [nowPageList, setNowPageList] = useState([]);
+//   const secondFilterList = ["봄", "여름", "가을", "겨울"];
+
+//   useEffect(() => {
+//     fetchSeasonList(secondFilterList[secondFilterIdx]).then((r) => {
+//       setSeasonClothList(r);
+//       setNowPageList(r.slice((page - 1) * 10, page * 10));
+
+//       // 마지막 페이지 번호 설정
+//       setMaxPage(Math.ceil(r.length / 10));
+//     });
+//   }, [page, secondFilterIdx]);
+
+//   return <SeasonPresenter clothList={nowPageList} />;
+// };
