+++ frontend/src/Presenter/Screens/Main/components/Total/TotalContainer.jsx
@@ -1,14 +1,22 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
+import { fetchClothList } from "../../../../../httpRequest";
 import TotalPresenter from "./TotalPresenter";
 
-const TotalContainer = ({ clothList, page, setMaxPage }) => {
-  //현재 보여지고 있는 의상 리스트
-  const nowPageClothList = clothList.slice((page - 1) * 10, page * 10);
+const TotalContainer = ({ page, setMaxPage }) => {
+  const [clothList, setClothList] = useState([]); // 전체 의상 리스트
+  const [nowPageList, setNowPageList] = useState([]); // 현재 보여지고 있는 페이지 리스트
 
-  // 마지막 페이지 번호 설정
-  setMaxPage(Math.ceil(clothList.length / 10));
+  useEffect(() => {
+    fetchClothList().then((r) => {
+      setClothList(r);
+      setNowPageList(r.slice((page - 1) * 10, page * 10));
 
-  return <TotalPresenter clothList={nowPageClothList} />;
+      // 마지막 페이지 번호 설정
+      setMaxPage(Math.ceil(r.length / 10));
+    });
+  }, [page]);
+
+  return <TotalPresenter clothList={nowPageList} />;
 };
 
 export default TotalContainer;
